build(packaging): migrate project from pup
Project needs own github pages; fork over and build for prod
