+++ 404.html
@@ -23,7 +23,7 @@
       // https://username.github.io/repo-name/one/two?a=b&c=d#qwe becomes
       // https://username.github.io/repo-name/?p=/one/two&q=a=b~and~c=d#qwe
       // Otherwise, leave segmentCount as 0.
-      var segmentCount = 1;
+      var segmentCount = 0;
 
       var l = window.location;
       l.replace(
