+++ index.html
@@ -1,4 +1,4 @@
-<!DOCTYPE html><html lang="en"><head><base href="/pup/"><title>Richard Q. Huynh, esq.</title><meta charset="utf-8"><meta http-equiv="X-UA-Compatible" content="IE=edge"><meta name="robots" content="noindex, nofollow"><meta name="viewport" content="width=device-width"><link href="https://fonts.googleapis.com/css?family=Open+Sans:400,600,700" rel="stylesheet"><script type="text/javascript">(function(l) {
+<!DOCTYPE html><html lang="en"><head><base href="/"><title>Richard Q. Huynh, esq.</title><meta charset="utf-8"><meta http-equiv="X-UA-Compatible" content="IE=edge"><meta name="robots" content="noindex, nofollow"><meta name="viewport" content="width=device-width"><link href="https://fonts.googleapis.com/css?family=Open+Sans:400,600,700" rel="stylesheet"><script type="text/javascript">(function(l) {
   if (l.search) {
     var q = {};
     l.search.slice(1).split('&').forEach(function(v) {
@@ -14,4 +14,4 @@
     }
   }
 }(window.location))
-</script></head><body ng-app="rqhlaw"><main ui-view></main><script type="text/javascript" src="/pup/app.bundle.js?7345fc042e44fd7b9eae"></script></body></html>
+</script></head><body ng-app="rqhlaw"><main ui-view></main><script type="text/javascript" src="/dist/app.bundle.js?6678c1fda9eec5ed720c"></script></body></html>
